+++ veritabaniProje/Entity/tSatis.cs
@@ -12,6 +12,6 @@ namespace veritabaniProje.Entity
         [Key]
         public int satisNo { get; set; }
         public string satisTuru { get; set; }
-        public double satisTutar { get; set; }
+        public float satisTutar { get; set; }
     }
 }
