+++ veritabaniProje/Entity/tTedarikci.cs
@@ -13,7 +13,7 @@ namespace veritabaniProje.Entity
         public int tedarikciId { get; set; }
         public string urunAdi { get; set; }
         public int urunId { get; set; }
-        public double borcMiktar { get; set; }
+        public float borcMiktar { get; set; }
         public int urunMiktar { get; set; }
 
     }
